theme: distinguish code better
I found it really hard to read what was actually code, so this change
adds a very subtle hint of colour around all the code which has cheeky
border radii. I've also cheekily made the font a tiny bit smaller for
inline code so that even with the required padding, it does not mess up
the line spacing visibly.