type: readjust font sizes and spacings a little
This is in response to a minor complaint about the text size on
mastodon, which I agree with:
https://chaos.social/@imadnyc@mstdn.social/112961604691513932

I personally agree that 20px does not feel right, and more to the point,
our previous type scale was really odd: body text was 1.25rem (20px),
while the nav links were 1rem (16px). It seems like the intent of this
was to make it so that the body text was not 16px, which felt too small
(and does to me too).

However, it seems to me like the *actual* solution here is to set the
entire type scale overall larger, then decrease the size of any type
that looks particularly silly.

Thus, I've set the root font-size to 18px (via a percentage of 112.5%)
instead of the default 16px, which will fix all the 1rem font sizes
elsewhere in the site, to *make* 1rem a reasonable body text size.