This adds dates to news posts in both the list and the post page.
It uses very simple styling; you might disagree.

In the list, the date is displayed below the title, with a bit of
padding below.

In the post, it's also displayed below the title, centered. If the
description is provided, it will have a little bit of padding, but so
far, no news posts have descriptions.